Update index.html
New version of DanIdle, built from the ground up (yes, again). This time I made use of many more Functional Programming methods, along with Object Composition for the blocks. I think this has become a good example of Object Composition in action; Blocks are built using add-on code blocks where needed. For example, any block that requires a tool gets the blockRequiresTool add-on added to it, that provides the code needed to show, gather and use tools