Completed fireminer
Fire Miner block is working... for now. We don't currently have any way to generate ores, but that shouldn't be too difficult to add. Before we can do that, a lot of map generator updates need to take place. I've been kind of putting off work on the map manager, but we can't do that anymore