BoulderWall & DirtMaker
Yes, there's been a lot break, but been doing other things. This adds a Boulder Wall and Dirt Maker. I also changed the way LastBlockId is used in each block; no need to have a shared variable that each block has to manage, when a function to handle that will do